# Improve performance of link transforms streaming protocol