--- a/src/meshcat_viz/__init__.py
+++ b/src/meshcat_viz/__init__.py
@@ -1,3 +1,4 @@
+from . import meshcat
 from .model import MeshcatModel
 from .model_builder import MeshcatModelBuilder
 from .world import MeshcatWorld