--- a/src/meshcat_viz/meshcat/__init__.py
+++ b/src/meshcat_viz/meshcat/__init__.py
@@ -1,0 +1,1 @@
+from . import commands, server, visualizer