| Model | Links | Pre | Post | Improvement |
| --- | :---: | :---: | :---: | :---: |
| [`panda`][panda] | 10 | 17.3 ms ± 1.02 ms | 4.92 ms ± 300 µs | x3.5 |
| [`icub`][icub] | 33 | 50.7 ms ± 2.36 ms | 6.19 ms ± 260 µs | x8.2 |

[icub]: https://github.com/robotology/gym-ignition-models/tree/088a81c55970ca5eec51108d249b8ce3dfd09e7a/src/gym_ignition_models/iCubGazeboV2_5
[panda]: https://github.com/robotology/gym-ignition-models/tree/088a81c55970ca5eec51108d249b8ce3dfd09e7a/src/gym_ignition_models/panda

<details>
<summary>Testing script</summary>

```python
import pathlib

import gym_ignition_models
import numpy as np

from meshcat_viz.world import MeshcatWorld

if __name__ == "__main__":

    model_sdf_path = pathlib.Path(
        gym_ignition_models.get_model_resource(
            # robot_name="iCubGazeboV2_5",
            robot_name="panda",
            resource_type=gym_ignition_models.ResourceType.SDF_PATH,
        )
    )

    # Open the visualizer
    world = MeshcatWorld()
    world.open()

    # Insert the model from a URDF/SDF resource.
    # Note: for URDF files support, check details in https://github.com/ami-iit/rod.
    model_name = world.insert_model(sdf=model_sdf_path)

    # Update the joint positions
    s = world._jaxsim_models[model_name].joint_random_positions()
    world.update_model(model_name=model_name, joint_positions=s)

    %timeit -r10 -n25 world.update_model(model_name=model_name, joint_positions=s)
```

</details>

Visualization in real-time has visually good performance in the range 30-60 fps (33 ms and 16 ms, respectively).
With this PR, also models with a lot of DoFs are compatible with such rate.